requirements.txt
pyyaml
pandas
openmeteo-requests
requests-cache
retry-requests
google-cloud-storage
pyarrow